trying to link against LLVM

+++ src/libponyc/ast/ast.c
@@ -306,7 +306,7 @@ double ast_float(ast_t* ast)
 
 size_t ast_int(ast_t* ast)
 {
-  return token_int(ast->t);
+  return (size_t)token_int(ast->t);
 }
 
 ast_t* ast_type(ast_t* ast)
